rewrite CameraController.java to work with both camera1 and camera2
