+++ TMessagesProj/src/main/java/org/telegram/messenger/camera/CameraSession.java
@@ -116,7 +116,7 @@ public class CameraSession {
     }
 
     public void checkFlashMode(String mode) {
-        ArrayList<String> modes = CameraController.getInstance().availableFlashModes;
+        ArrayList<String> modes = CameraController.getInstance().getAvailableFlashModes();
         if (modes.contains(currentFlashMode)) {
             return;
         }
@@ -147,7 +147,7 @@ public class CameraSession {
     }
 
     public String getNextFlashMode() {
-        ArrayList<String> modes = CameraController.getInstance().availableFlashModes;
+        ArrayList<String> modes = CameraController.getInstance().getAvailableFlashModes();
         for (int a = 0; a < modes.size(); a++) {
             String mode = modes.get(a);
             if (mode.equals(currentFlashMode)) {
